requirements.txt
pytorch
torchvision
scikit-image
matplotlib
nb_conda
ipykernel
numpy
